requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
Django==3.0.2
djangorestframework==3.11.0
idna==2.8
pdfkit==0.6.1
pytz==2019.3
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.7
